% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_icons.R
\name{game_icon}
\alias{game_icon}
\title{Create a game icon}
\usage{
game_icon(icon, class = NULL, style = NULL)
}
\arguments{
\item{icon}{name of the icon}

\item{class}{Additional classes to customize the style of the icon.}
}
\description{
Create a game icon
}
