% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_icons.R
\name{list_icons}
\alias{list_icons}
\title{List all available icons}
\usage{
list_icons()
}
\description{
Lists all available icons. Requires stringr which is only listed in Enchances
so it might not be installed
}
